+ small bug fix in delete(Long id) method (ProductDAO)

+++ internetShop/src/main/java/lv/javaguru/java2/database/hibernate/ProductDAOImpl.java
@@ -39,7 +39,8 @@ public class ProductDAOImpl implements ProductDAO {
     public void delete(Long id) throws DBException {
         Session session = sessionFactory.getCurrentSession();
         Product product = (Product) session.get(Product.class, id);
-        session.delete(product);
+        if (product != null)
+            session.delete(product);
     }
 
     @Override
